+++ example/tests/conftest.py
@@ -1,10 +1,12 @@
 import pytest
 from pytest_factoryboy import register
+from rest_framework.test import APIClient
 
 from example.factories import (
     ArtProjectFactory,
     AuthorBioFactory,
     AuthorFactory,
+    AuthorTypeFactory,
     BlogFactory,
     CommentFactory,
     CompanyFactory,
@@ -16,6 +18,7 @@ from example.factories import (
 register(BlogFactory)
 register(AuthorFactory)
 register(AuthorBioFactory)
+register(AuthorTypeFactory)
 register(EntryFactory)
 register(CommentFactory)
 register(TaggedItemFactory)
@@ -54,3 +57,8 @@ def single_company(art_project_factory, research_project_factory, company_factor
 @pytest.fixture
 def single_art_project(art_project_factory):
     return art_project_factory()
+
+
+@pytest.fixture
+def client():
+    return APIClient()